Add files via upload
Bridge turning into a sensor for some reason keeps getting shoved and moved by the player. I've tried to make it stay still for a few frame after turning into a sensor, and tried keeping it static a little after turning into a sensor, so that player would not collide with it while falling. Didn't work
Workaround: make new FBox, with bridge sprite, that is set to sensor and not static in it's setup. this worked for some reason
added slow falling to bridge
added shaking and "hp" to bridge (player can touch it for 25 frames before it falls, the frame count is paused when player stops touching it)
adjusted lava animation (+ randomness when bubble)
adjusted water animation
adjusted/edited many sprites, and made a new lava sprite to slow more smoothly
FGameObject made
Fixed lag when camera is following player
Added (and make a couple more) sprites for player animation
Player animation started, and being finished soon
Enemies... not quite started yet, but I've watched the vids and know what to do for tmr
^Includes all work done since last upload to GitHub (I forgot to upload yesterday)